loader - move progress-with-label to its own module
